Implement P² + K² + P×K matrix storage architecture in RAPTOR engine
- Add K×K inter-centroid distance matrix with upper triangle storage (87.5% compression)
- Implement P×K vector-to-centroid matrices with adaptive strategies:
  - Full storage when K ≤ D (dimension)
  - Hierarchical storage (mean + sparse deltas) when K ≤ 10D (99.85% compression)
  - Sparse storage (top-√D centroids) when K > 10D
- Use 16-bit quantization for all distance matrices (50% additional savings)
- Update writer to build and serialize matrices in footer
- Add compute_all_centroid_distances method for K×K matrix construction
- Integrate matrix building into finalize() method
- Zero-copy design with index-based access (no ID storage overhead)

+++ src/storage/engines/raptor/consolidated_compactor.rs
@@ -304,7 +304,7 @@ impl RaptorCompactor {
         let mut current_offset = 0u64;
         
         for (idx, chunk) in vectors.chunks(group_size).enumerate() {
-            let mut row_group = RowGroup::new(idx as u32);
+            let mut row_group = RowGroup::new(idx as u16);
             row_group.row_count = chunk.len();
             row_group.offset = current_offset;
             row_group.vectors = Some(chunk.to_vec());
